Put the states on specific components rather than all on log info page.
